FIX : compile warning
git-svn-id: http://svn.openstreetmap.org/applications/editors/merkaartor@10331 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ Map/Coord.h
@@ -34,7 +34,7 @@ inline double intToRad(int a)
 	return double(a)*M_PI/INT_MAX;
 }
 
-inline double radToInt(double x)
+inline int radToInt(double x)
 {
 	return static_cast<int>(x*INT_MAX/M_PI);
 }
